// web/src/shared/api/generated/model/sessionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { JsonElement } from './jsonElement';
import type { SessionItemResult } from './sessionItemResult';

export interface SessionResult {
  id: string;
  kind: string;
  scope: JsonElement;
  isTestMode: boolean;
  startedAt: string;
  /** @nullable */
  submittedAt: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  scorePercent: number | string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  timeTakenMilliseconds: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  currentPosition: number | string | null;
  items: SessionItemResult[];
}